Do not remove backup if org does not exist

+++ UABEAvalonia/Forms/MainWindow.axaml.cs
@@ -787,7 +787,7 @@ namespace UABEAvalonia
                 await OpenFiles([filePath]);
             }
 
-            if (!string.IsNullOrWhiteSpace(backupPath))
+            if (!string.IsNullOrWhiteSpace(backupPath) && File.Exists(filePath))
             {
                 await CloseAllFiles();
                 File.Delete(backupPath);
